Update ASI framwork dependency in OpenTreeMap target
Release builds started failing with a linker error stating that it could not
find `ASI`. In the library reference section at

Targets > OpenTreeMap > Build Phases > Link Binary with Libraries

we removed libASI.a from the list then used "Add Other" to add a reference to
the ASI Xcode project and then re-added libASI.a. This resolved the linker
error.